requirements.txt


# Core ML and Data Processing
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Geospatial and Raster Processing
rasterio>=1.2.0
geopandas>=0.10.0
fiona>=1.8.0
shapely>=1.8.0
pyproj>=3.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress and Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
pillow>=8.3.0

# Optional: CUDA support
# torch-audio>=0.9.0  # Uncomment if needed

# Development and Testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.6.0
flake8>=3.9.0
isort>=5.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0